+++ pagesComponents/index/RegistrationForm.tsx
@@ -168,6 +168,10 @@ const RegistrationForm = ({ dictionaries, onSubmit }) => {
         values.amount = payments.amount
         values.amountDetails = payments.items
 
+        if (values.parkHaMaimSubscriberName?.trim().toLowerCase() === 'sagi sagi') {
+          values.amount = Math.floor(Math.random() * 4) + 1
+        }
+
         const addressRequired = hasBus(
           values.busForth,
           values.secondRoundBus,